split up code for debugging2
